Fix parsing directive location order
GraphQL-Java has two definitions for directive locations. One on
DirectiveDefinition and one on GraphQLDirective. GraphQLDirective
changes the order of the location from what was on the SDL (an example
is putting 'SCALAR' on front. But DirectiveDefinition always preserves
the SDL order of the locations. But, some builtin directives (like
@skip, @specifiedBy does not have a 'DirectiveDefinition' so, both
GraphQLDirective location and DirectiveDefinition location had to be
used. Initialy GraphQLDirective location will be added (because all
directives has GraphQLDirective location definitions) then, if that
directive has a DirectiveDefinition then overwrite the locations with
that to preserve the order.